requirements.txt
httpx>=0.25
playwright>=1.40
python-dotenv>=1.0
fastapi>=0.110
uvicorn[standard]>=0.27
tzdata>=2024.1

